weather/requirements.txt
apache-airflow-providers-google
pandas
pandas-gbq
google-cloud-bigquery
dbt-bigquery
requests